refactor(chat): extract socket event wiring into hook
