+++ src/pages/Chat.jsx
@@ -17,13 +17,12 @@ import ConversationList from "../components/chat/ConversationList";
 import ConversationHeader from "../components/chat/ConversationHeader";
 import ArchivedTeamBanner from "../components/chat/ArchivedTeamBanner";
 import MessageDisplay from "../components/chat/MessageDisplay";
-import { parseSystemMessage } from "../utils/messageSystemParser";
 import MessageInput from "../components/chat/MessageInput";
 import { useAuth } from "../contexts/AuthContext";
 import { messageService } from "../services/messageService";
 import socketService from "../services/socketService";
-import useSocketEvents from "../hooks/useSocketEvents";
 import useChatTyping from "../hooks/useChatTyping";
+import useChatSocketEvents from "../hooks/useChatSocketEvents";
 import { userService } from "../services/userService";
 import { isQuietError } from "../services/api";
 import { teamService } from "../services/teamService";
@@ -43,17 +42,13 @@ import {
   formatArchiveTimeRemaining,
   msUntilNextArchiveChange,
 } from "../utils/dateHelpers";
-import { getMessageConversationTarget } from "../utils/messageNotificationUtils";
 import {
   getConversationPartnerId,
   isActiveTeamMemberRow,
   isUserTeamMember,
-  getPayloadTeamId,
-  isCurrentUserRemovalPayload,
   mergeTeamDetailsIntoConversationData,
   isArchivedTeamData,
   getConversationUpdatedAt,
-  isDirectConversationForPartner,
 } from "../utils/chatHelpers";
 import {
   CHAT_SEARCH_PAGE_SIZE,
@@ -1289,525 +1284,29 @@ const Chat = () => {
     pendingScrollAdjustmentRef.current = null;
   }, [messages, hasMoreMessages]);
 
-  const handleKickedFromTeam = useCallback(
-    (data) => {
-      if (
-        conversationType === "team" &&
-        parseInt(conversationId, 10) === data.teamId
-      ) {
-        revokeTeamChatAccess(data.teamId, "You have been removed from this team.");
-        return;
-      }
-
-      setConversations((prev) =>
-        prev.filter((c) => !(c.type === "team" && c.id === data.teamId)),
-      );
-    },
-    [conversationId, conversationType, revokeTeamChatAccess, setConversations],
-  );
-
-  // Set up WebSocket event listeners
-  useSocketEvents((socket) => {
-    if (!socket || !isAuthenticated) {
-      return undefined;
-    }
-
-    // Handle online users
-    const handleOnlineUsers = (users) => {
-      setOnlineUsers(users);
-    };
-
-    // Handle new messages
-    const handleNewMessage = (message) => {
-      const messageTarget = getMessageConversationTarget(message, user?.id);
-      const messageConvId = String(
-        messageTarget.conversationId ?? message.conversationId,
-      );
-      const currentConvId = String(conversationId);
-
-      // Get current conversation type from URL
-      const urlParams = new URLSearchParams(window.location.search);
-      const currentType = urlParams.get("type") || "direct";
-
-      // Check if message belongs to current conversation
-      let isForCurrentConversation = false;
-
-      if (message.type === currentType) {
-        if (currentType === "team") {
-          // For team chats: conversationId must match
-          isForCurrentConversation = messageConvId === currentConvId;
-        } else {
-          // For DMs: either I sent it to this person, or this person sent it to me
-          const isSentByMe =
-            message.senderId === user?.id && messageConvId === currentConvId;
-          const isReceivedFromThem = String(message.senderId) === currentConvId;
-          isForCurrentConversation = isSentByMe || isReceivedFromThem;
-        }
-      }
-
-      if (isForCurrentConversation) {
-        if (currentType === "team") {
-          const parsedMessage = parseSystemMessage(message.content);
-          const removedMemberId =
-            parsedMessage?.memberId ?? parsedMessage?.userId ?? null;
-
-          if (
-            ["member_removed", "member_removed_public"].includes(parsedMessage?.type) &&
-            removedMemberId != null &&
-            String(removedMemberId) === String(user?.id)
-          ) {
-            revokeTeamChatAccess(
-              messageConvId,
-              "You have been removed from this team.",
-            );
-            return;
-          }
-        }
-
-        setMessages((prev) => {
-          // If this is our own message, replace the optimistic version
-          if (message.senderId === user.id) {
-            const withoutOptimistic = prev.filter(
-              (msg) => !msg.isOptimistic || msg.senderId !== user.id,
-            );
-
-            const messageExists = withoutOptimistic.some(
-              (msg) => msg.id === message.id,
-            );
-            if (messageExists) {
-              return prev;
-            }
-
-            const newMessage = {
-              id: message.id,
-              senderId: message.senderId,
-              content: message.content,
-              imageUrl: message.imageUrl,
-              fileUrl: message.fileUrl,
-              fileName: message.fileName,
-              createdAt: message.createdAt,
-              senderUsername: message.senderUsername,
-              type: message.type,
-              fileSize: message.fileSize,
-              fileExpiresAt: message.fileExpiresAt,
-              fileDeletedAt: message.fileDeletedAt,
-              readCount: message.readCount,
-              recipientCount: message.recipientCount,
-              readByUsers: message.readByUsers,
-              editedAt: message.editedAt || message.edited_at,
-              editedBy: message.editedBy ?? message.edited_by,
-              isEdited: message.isEdited || message.is_edited,
-              replyTo: message.replyTo,
-              replyToId: message.replyToId || message.reply_to_id,
-            };
-
-            return dedupeMessages([...withoutOptimistic, newMessage]);
-          } else {
-            const messageExists = prev.some((msg) => msg.id === message.id);
-            if (messageExists) {
-              return prev;
-            }
-
-            const newMessage = {
-              id: message.id,
-              senderId: message.senderId,
-              content: message.content,
-              imageUrl: message.imageUrl,
-              fileUrl: message.fileUrl,
-              fileName: message.fileName,
-              createdAt: message.createdAt,
-              senderUsername: message.senderUsername,
-              type: message.type,
-              fileSize: message.fileSize,
-              fileExpiresAt: message.fileExpiresAt,
-              fileDeletedAt: message.fileDeletedAt,
-              readCount: message.readCount,
-              recipientCount: message.recipientCount,
-              readByUsers: message.readByUsers,
-              editedAt: message.editedAt || message.edited_at,
-              editedBy: message.editedBy ?? message.edited_by,
-              isEdited: message.isEdited || message.is_edited,
-              replyTo: message.replyTo,
-              replyToId: message.replyToId || message.reply_to_id,
-            };
-
-            return dedupeMessages([...prev, newMessage]);
-          }
-        });
-
-        // Mark as read if viewing and didn't send it
-        if (message.senderId !== user.id) {
-          const urlParams = new URLSearchParams(window.location.search);
-          const type = urlParams.get("type") || "direct";
-          socketService.markMessagesAsRead(currentConvId, type);
-        }
-      }
-
-      // Update conversation list
-      setConversations((prev) => {
-        const isUnreadIncoming =
-          messageConvId !== currentConvId && message.senderId !== user.id;
-        const lastMessagePreview = buildConversationLastMessagePreview(message);
-        let conversationUpdated = false;
-        const updatedList = prev.map((conv) => {
-          if (String(conv.id) === messageConvId) {
-            conversationUpdated = true;
-            const currentUnreadCount = conv.unreadCount ?? conv.unread_count ?? 0;
-            const unreadCount = isUnreadIncoming
-              ? currentUnreadCount + 1
-              : currentUnreadCount;
-
-            return {
-              ...conv,
-              lastMessage: lastMessagePreview,
-              updatedAt: message.createdAt,
-              isVirtual: false,
-              unreadCount,
-              unread_count: unreadCount,
-            };
-          }
-          return conv;
-        });
-
-        if (!conversationUpdated) {
-          if (messageTarget.type === "direct" && message.senderId !== user.id) {
-            updatedList.unshift({
-              id: Number.isNaN(Number(messageConvId))
-                ? messageConvId
-                : Number(messageConvId),
-              type: "direct",
-              partner: {
-                id: message.senderId,
-                username: message.senderUsername,
-                firstName: message.senderFirstName,
-                lastName: message.senderLastName,
-                avatarUrl: message.senderAvatarUrl,
-              },
-              lastMessage: lastMessagePreview,
-              updatedAt: message.createdAt,
-              unreadCount: isUnreadIncoming ? 1 : 0,
-              unread_count: isUnreadIncoming ? 1 : 0,
-            });
-          }
-
-          refreshConversationList();
-        }
-
-        const deduplicatedList = dedupeConversations(updatedList);
-
-        return deduplicatedList.sort((a, b) => {
-          const aDate = normalizeTimestampToDate(a.updatedAt)?.getTime() ?? 0;
-          const bDate = normalizeTimestampToDate(b.updatedAt)?.getTime() ?? 0;
-          return bDate - aDate;
-        });
-      });
-    };
-
-    const refreshTeamEventMessages = async (payload) => {
-      const urlParams = new URLSearchParams(window.location.search);
-      const currentType = urlParams.get("type") || "direct";
-      const activeTeamId =
-        activeConversationRef.current?.team?.id ?? activeConversationRef.current?.id;
-      const teamId =
-        getPayloadTeamId(payload) ??
-        (currentType === "team" ? activeTeamId : null);
-      if (!teamId) return;
-
-      if (isCurrentUserRemovalPayload(payload, user?.id)) {
-        revokeTeamChatAccess(teamId, "You have been removed from this team.");
-        return;
-      }
-
-      refreshConversationList();
-
-      if (
-        currentType !== "team" ||
-        String(teamId) !== String(conversationId)
-      ) {
-        return;
-      }
-
-      try {
-        const canStillAccess = await refreshActiveTeamMembership(teamId);
-
-        if (!canStillAccess) {
-          return;
-        }
-
-        const messagesResponse = await messageService.getMessages(
-          teamId,
-          "team",
-        );
-        const fetchedMessages = messagesResponse.data || [];
-        setHasMoreMessages(messagesResponse.hasMore || false);
-        setMessages(dedupeMessages(fetchedMessages));
-        socketService.markMessagesAsRead(teamId, "team");
-      } catch (err) {
-        console.error("Error refreshing team event messages:", err);
-      }
-    };
-
-    // Handle message status updates
-    const handleMessageStatus = (data) => {
-      if (String(data.conversationId) === String(conversationId)) {
-        const readCountByMessageId = new Map(
-          (data.messageReadCounts || []).map((status) => [
-            String(status.messageId),
-            status,
-          ]),
-        );
-
-        setMessages((prev) =>
-          prev.map((msg) => {
-            if (data.type === "team") {
-              const status = readCountByMessageId.get(String(msg.id));
-
-              if (!status) {
-                return msg;
-              }
-
-              return {
-                ...msg,
-                readAt: msg.readAt || status.firstReadAt || data.readAt,
-                readCount: status.readCount,
-                recipientCount: status.recipientCount,
-                readByUsers: status.readByUsers || msg.readByUsers,
-              };
-            }
-
-            if (msg.senderId !== user.id) {
-              return msg;
-            }
-
-            return {
-              ...msg,
-              readAt: msg.readAt || data.readAt,
-              readCount: 1,
-              recipientCount: 1,
-            };
-          }),
-        );
-      }
-    };
-
-    // Handle conversation updates
-    const handleConversationUpdate = (data) => {
-      setConversations((prev) => {
-        const conversationIndex = prev.findIndex(
-          (c) => String(c.id) === String(data.id),
-        );
-
-        if (conversationIndex === -1) {
-          refreshConversationList();
-          return prev;
-        }
-
-        const updatedList = [...prev];
-        updatedList[conversationIndex] = {
-          ...updatedList[conversationIndex],
-          lastMessage:
-            buildConversationLastMessagePreview(data) ||
-            data.lastMessage ||
-            updatedList[conversationIndex].lastMessage,
-          updatedAt: data.updatedAt,
-        };
-
-        const deduplicatedList = dedupeConversations(updatedList);
-
-        return deduplicatedList.sort((a, b) => {
-          const aDate = normalizeTimestampToDate(a.updatedAt)?.getTime() ?? 0;
-          const bDate = normalizeTimestampToDate(b.updatedAt)?.getTime() ?? 0;
-          return bDate - aDate;
-        });
-      });
-    };
-
-    const handleTeamMemberLeft = (data) => {
-      if (!data?.teamId) return;
-      const leftUserId = data.userId ?? data.user_id ?? null;
-
-      setTeamMembersRefreshSignal({
-        teamId: data.teamId,
-        userId: leftUserId,
-        receivedAt: Date.now(),
-      });
-
-      const hasTeamConversation = conversationsRef.current.some(
-        (conversation) =>
-          conversation.type === "team" &&
-          String(conversation.id) === String(data.teamId),
-      );
-
-      if (hasTeamConversation) {
-        refreshConversationList();
-      }
-
-      const activeTeamId =
-        activeConversationRef.current?.team?.id ?? activeConversationRef.current?.id;
-
-      if (String(activeTeamId) !== String(data.teamId)) {
-        return;
-      }
-
-      if (leftUserId != null && String(leftUserId) === String(user?.id)) {
-        revokeTeamChatAccess(data.teamId, "You have left this team chat.");
-        return;
-      }
-
-      refreshActiveTeamMembership(data.teamId)
-        .catch((err) =>
-          console.error("Error refreshing active team members:", err),
-        );
-    };
-
-    const handleConversationDeleted = (data) => {
-      if (!data?.partnerId) return;
-
-      setConversations((prev) =>
-        prev.filter(
-          (conversation) =>
-            !isDirectConversationForPartner(conversation, data.partnerId),
-        ),
-      );
-
-      const currentUrlType =
-        new URLSearchParams(window.location.search).get("type") ||
-        activeConversationRef.current?.type ||
-        "direct";
-
-      const activePartnerId =
-        getConversationPartnerId(activeConversationRef.current) ??
-        getConversationPartnerId(
-          conversationsRef.current.find(
-            (conversation) =>
-              conversation.type === "direct" &&
-              String(conversation.id) === String(conversationId),
-          ),
-        );
-
-      const isCurrentConversationDeleted =
-        currentUrlType === "direct" &&
-        (String(activePartnerId ?? "") === String(data.partnerId) ||
-          String(conversationId ?? "") === String(data.partnerId));
-
-      if (!isCurrentConversationDeleted) {
-        return;
-      }
-
-      setError("This conversation is no longer available.");
-      setActiveConversation(null);
-      setMessages([]);
-      clearTypingUsers();
-      setHighlightMessageIds([]);
-      setHasMoreMessages(false);
-      navigate("/chat", { replace: true });
-    };
-
-    // Handle message deleted (soft delete broadcast)
-    const handleMessageDeleted = (payload) => {
-      // payload: { messageId, deletedAt, deletedBy, type, teamId, senderId, receiverId }
-      setMessages((prev) =>
-        prev.map((m) =>
-          String(m.id) === String(payload.messageId)
-            ? {
-                ...m,
-                deletedAt: payload.deletedAt || new Date().toISOString(),
-                deletedBy: payload.deletedBy,
-                content: null,
-                imageUrl: null,
-                fileUrl: null,
-                fileName: null,
-                fileSize: null,
-              }
-            : m,
-        ),
-      );
-      refreshConversationList();
-    };
-
-    const handleMessageEdited = (payload) => {
-      const messageId = payload.messageId ?? payload.id;
-      if (!messageId) return;
-      const nextContent = payload.content;
-
-      setMessages((prev) =>
-        prev.map((m) =>
-          String(m.id) === String(messageId)
-            ? {
-                ...m,
-                content: nextContent ?? m.content,
-                editedAt:
-                  payload.editedAt ||
-                  payload.edited_at ||
-                  payload.updatedAt ||
-                  payload.updated_at ||
-                  new Date().toISOString(),
-                editedBy: payload.editedBy ?? payload.edited_by ?? m.editedBy,
-                isEdited: true,
-              }
-            : m,
-        ),
-      );
-
-      if (payload.isLatestMessage || payload.isLastMessage) {
-        setConversations((prev) =>
-          prev.map((conversation) =>
-            String(conversation.id) === String(payload.conversationId) ||
-            (payload.type === "direct" &&
-              (String(conversation.id) === String(payload.senderId) ||
-                String(conversation.id) === String(payload.receiverId)))
-              ? {
-                  ...conversation,
-                  lastMessage: nextContent ?? conversation.lastMessage,
-                }
-              : conversation,
-          ),
-        );
-      }
-    };
-
-    // Subscribe to events
-    socket.on("users:online", handleOnlineUsers);
-    socket.on("message:received", handleNewMessage);
-    socket.on("typing:update", handleTypingUpdate);
-    socket.on("message:status", handleMessageStatus);
-    socket.on("conversation:updated", handleConversationUpdate);
-    socket.on("team:member_left", handleTeamMemberLeft);
-    socket.on("conversation:deleted", handleConversationDeleted);
-    socket.on("team:member_kicked", handleKickedFromTeam);
-    socket.on("message:deleted", handleMessageDeleted);
-    socket.on("message:edited", handleMessageEdited);
-    socket.on("notification:new", refreshTeamEventMessages);
-
-    // Cleanup function to remove listeners
-    return () => {
-      socket.off("users:online", handleOnlineUsers);
-      socket.off("message:received", handleNewMessage);
-      socket.off("typing:update", handleTypingUpdate);
-      socket.off("message:status", handleMessageStatus);
-      socket.off("conversation:updated", handleConversationUpdate);
-      socket.off("team:member_left", handleTeamMemberLeft);
-      socket.off("conversation:deleted", handleConversationDeleted);
-      socket.off("team:member_kicked", handleKickedFromTeam);
-      socket.off("message:deleted", handleMessageDeleted);
-      socket.off("message:edited", handleMessageEdited);
-      socket.off("notification:new", refreshTeamEventMessages);
-    };
-  }, [
+  useChatSocketEvents({
+    activeConversationRef,
+    clearTypingUsers,
     conversationId,
-    handleKickedFromTeam,
-    fetchTeamDetails,
+    conversationType,
+    conversationsRef,
+    dedupeMessages,
+    handleTypingUpdate,
     isAuthenticated,
     navigate,
     refreshActiveTeamMembership,
     refreshConversationList,
     revokeTeamChatAccess,
-    activeConversation,
-    handleTypingUpdate,
-    searchParams,
-    user?.id,
-    user?.username,
-  ]);
+    setActiveConversation,
+    setConversations,
+    setError,
+    setHasMoreMessages,
+    setHighlightMessageIds,
+    setMessages,
+    setOnlineUsers,
+    setTeamMembersRefreshSignal,
+    user,
+  });
 
   const handleHeaderTeamClick = (e) => {
     e.stopPropagation();
